Move TS hooks types into ./versionedTypes
