+++ src/query/react/versionedTypes/index.ts
@@ -1,0 +1,1 @@
+export { TS41Hooks } from './ts41Types'
