+++ src/query/react/versionedTypes/package.json
@@ -1,0 +1,15 @@
+{
+  "types": "./index.d.ts",
+  "typesVersion": {
+    ">=4.1": {
+      "index.d.ts": [
+        "./ts41Types.d.ts"
+      ]
+    },
+    "<4.1": {
+      "index.d.ts": [
+        "./ts40Types.d.ts"
+      ]
+    }
+  }
+}
